Keep global-mutation security proofs at the integration boundary
The intrinsic hardening regressions replace process-wide runtime globals, so place them in the integration suite where those effects are isolated and explicitly permitted.

Constraint: The semantic unit-boundary audit rejects process-global mutation in colocated unit tests.

Rejected: Add semantic migration dispositions | that would grow test-boundary debt for newly introduced coverage.

Confidence: high

Scope-risk: narrow

Directive: Tests that replace runtime globals belong under tests/integration, not colocated unit suites.

Tested: integration and affected unit tests, deno check, formatting, and semantic unit-boundary audit.

Not-tested: remote CI pending this commit.
